% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monad.R
\name{applicative}
\alias{applicative}
\alias{pure}
\alias{pure_like}
\alias{pure_like_}
\alias{ap}
\alias{ap_}
\alias{\%<*>\%}
\title{Applicative typeclass}
\usage{
pure(...)

pure_like(...)

pure_like_(...)

ap(...)

ap_(.f, .m, ...)

... \%<*>\% NA
}
\description{
Applicative typeclass
}
