% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monad.R
\name{functor}
\alias{functor}
\alias{fmap}
\alias{fmap_}
\alias{\%$\%}
\alias{\%&\%}
\alias{\%<$>\%}
\title{Functor typeclass}
\usage{
fmap(...)

fmap_(.f, .m, ...)

f \%$\% x

x \%&\% f

... \%<$>\% NA
}
\description{
Functor typeclass
}
