% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maybe.R
\docType{data}
\name{maybe}
\alias{maybe}
\alias{nothing}
\alias{just}
\title{Maybe monad}
\format{An object of class \code{nothing} (inherits from \code{maybe}, \code{monad}, \code{applicative}, \code{functor}) of length 0.}
\usage{
nothing

just(.v)
}
\description{
Maybe monad
}
\keyword{datasets}
