% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monad.R
\name{monad}
\alias{monad}
\alias{bind}
\alias{bind_}
\alias{\%>>=\%}
\alias{\%=<<\%}
\title{Monad typeclass}
\usage{
bind(.f, .m, ...)

bind_(.f, .m, ...)

m \%>>=\% f

... \%=<<\% NA
}
\description{
Monad typeclass
}
