% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rattlr.R
\name{python_eval}
\alias{python_eval}
\title{Evaluate expressions and/or statements in Python}
\usage{
python_eval(pc, exprs = c(), imports = c(), envir = NULL)
}
\description{
Evaluate expressions and/or statements in Python
}
