% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rattlr.R
\name{python_eval_assign}
\alias{python_eval_assign}
\title{Evaluate expressions and/or statements in Python and assign bindings}
\usage{
python_eval_assign(pc, exprs = c(), imports = c(), envir = globalenv())
}
\description{
Evaluate expressions and/or statements in Python and assign bindings
}
