\name{rattlr-package}
\alias{rattlr-package}
\alias{rattlr}
\docType{package}
\title{
\packageTitle{rattlr}
}
\description{
\packageDescription{rattlr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rattlr}
\packageIndices{rattlr}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{rattlr}

Maintainer: \packageMaintainer{rattlr}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
